Change initialization of Hankel transform
In some cases, the matrix representing the inverse Hankel Transform
is singular, and needs special treatment if it is to be inverted.

The previous treatment ensured back and forth retrieval (to machine
precision), but introduced spurious oscillations in the simulations.
The new treatment removes these oscillations, but back and forth
retrieval is not ensured anymore.