Fixed a bug in MPI + moving window
The bug was due to the fact that each MPI process was deciding
by how many cells to shift the window. This decision was based on the
value of `int((self.zmin - zmin_global)/dz)`, which was very sensitive to
machine precision errors (because the result is often very close to 1).

This lead some processors to move the window by a different amount then
the other processors. The solution adopted now is to have one processor
decide how much to move the window, and to broadcast that information to
all processors.