Implement cubic particle shapes on the CPU
This commit implements cubic particle shapes for the CPU. Keep in mind that cubic shapes are computationally very demanding. Since the new, more general routines used are strictly equivalent for linear shapes compared to the old routines, the latter have been removed entirely. In principle this implementation can be expanded to arbitrary order shapes, if the weights function is modified properly. This commit contains rountines for deposition and gathering.