Refactoring of exchange_fields
The exchange of fields between MPI domains is now organized in a more general way, so that arbitrary vector or scalar fields that are sent via MPI can either "replace" the domain fields or can be "added" to the domain fields.

Moreover, the GPU buffers will now be copied to pagelocked ("pinned") CPU arrays.